requirements.txt
Flask
Flask-Cors
psycopg2
gunicorn
python-dotenv
